+++ src/i18n/en/index.ts
@@ -23,7 +23,31 @@ const en = {
 	Password: 'Password',
 	LogIn: 'Log in',
 	UsernameExample: 'Jackson',
-	LoggedIn: '{username:string} Logged In'
+	LoggedIn: '{username:string} Logged In',
+	Card1Title: 'This is a title',
+	Card1Author: 'By Author',
+	Card1Description:
+		'Incididunt elit ipsum nulla pariatur dolore. Magna reprehenderit in nulla enim ad labore irure cillum',
+	Card1ExOne: 'Published',
+	Card1ExOneTime: '12st June, 2020',
+	Card1ExTwo: 'Reading time',
+	Card1ExTwoTime: '3 minute',
+
+	Card2TitleOne: 'Developer',
+	Card2TitleTwo: 'Name',
+	Card2Description:
+		'Esse fugiat ea eu velit dolore Lorem ullamco ex exercitation eu magna et do eiusmod. In qui mollit mollit exercitation quis fugiat commodo ullamco labore incididunt nulla eu officia. Dolore in culpa dolore ullamco veniam. Nostrud commodo eiusmod sint ipsum elit eiusmod. Qui adipisicing consectetur labore cupidatat quis aliquip veniam ut nostrud nulla commodo nisi. Est mollit laboris magna incididunt. Culpa cillum sit tempor consequat eu',
+
+	ShopCardPriceOne: 'Price',
+	ShopCardPriceTwo: '400,000',
+	ShopCardItemName: 'Food Name',
+	ShopCardItemOneTitle: 'Reserved',
+	ShopCardItemOneText: '2 tables',
+	ShopCardItemTwoTitle: 'Dessert',
+	ShopCardItemTwoText: '4 foods',
+	ShopCardItemThreeTitle: 'Drink',
+	ShopCardItemThreeText: '4 sodas',
+	ShopCardDescription: 'chicken barbecue'
 } satisfies BaseTranslation;
 
 export default en;