+++ src/i18n/fa/index.ts
@@ -23,7 +23,31 @@ const fa = {
 	Password: 'رمز عبور',
 	LogIn: 'ورود',
 	UsernameExample: 'محمد',
-	LoggedIn: '{username} وارد شد'
+	LoggedIn: '{username} وارد شد',
+	Card1Title: 'این سرتیتر است',
+	Card1Author: 'نویسنده',
+	Card1Description:
+		'لورم ایپسوم متن ساختگی با تولید سادگی نامفهوم از صنعت چاپ، و با استفاده از طراحان گرافیک است، چاپگرها و متون است.',
+	Card1ExOne: 'انتشارشده',
+	Card1ExOneTime: 'یکشنبه ،۲۲ دی ۱۳۹۸',
+	Card1ExTwo: 'زمان خواندن',
+	Card1ExTwoTime: '۳ دقیقه',
+
+	Card2TitleOne: 'توسعه دهنده',
+	Card2TitleTwo: 'اسم',
+	Card2Description:
+		'لورم ایپسوم متن ساختگی با تولید سادگی نامفهوم از صنعت چاپ، و با استفاده از طراحان گرافیک است، چاپگرها و متون بلکه روزنامه و مجله در ستون و سطرآنچنان که لازم است، و برای شرایط فعلی تکنولوژی مورد نیاز، و کاربردهای متنوع با هدف بهبود ابزارهای کاربردی می باشد، کتابهای زیادی در شصت و سه درصد گذشته حال و آینده، شناخت فراوان جامعه و متخصصان را می طلبد، تا با نرم افزارها شناخت بیشتری را برای طراحان رایانه ای علی الخصوص طراحان خلاقی، و فرهنگ پیشرو در زبان .',
+
+	ShopCardPriceOne: 'قیمت',
+	ShopCardPriceTwo: '400,000',
+	ShopCardItemName: 'غذا',
+	ShopCardItemOneTitle: 'سروشده',
+	ShopCardItemOneText: '۲ میز',
+	ShopCardItemTwoTitle: 'دسر',
+	ShopCardItemTwoText: '۴ غذا',
+	ShopCardItemThreeTitle: 'نوشیدنی',
+	ShopCardItemThreeText: '۴ نوشابه',
+	ShopCardDescription: 'جوجه کباب'
 } satisfies Translation;
 
 export default fa;